Identifying the CreditMinging tasks
The CreditMining tasks can now be identified.
The boosting manager is no longer considering any torrent which is not added by the manager itself.
(problem left that if the torrent is stopped by the manager, it will also no be considered in the future)
Moreover, when adding a new download, if it is being mined, it will be move out of the boosting manager and become a normal download.